# Add memory padding and alignment to prevent false sharing