Ensure no thirdparty thread writes to a cacheline shared by our ringbuffer, readIndex and writeIndex.
Ensure readIndex and writeIndex are located on different cachelines.