@nbronson 2016-04-13
@rigtorp Was there a particular use case where this made a difference?