@rigtorp 2016-04-13
False sharing between readIndex and writeIndex kills performance, this is a
small space trade off (~256 bytes) for a big performance gain. The cache
line ping-ponging also affects performance of unrelated threads, due to
memory bus traffic.

On Wed, Apr 13, 2016 at 9:28 AM Nathan Grasso Bronson <
notifications@github.com> wrote:

> @rigtorp <https://github.com/rigtorp> Was there a particular use case
> where this made a difference?
>
> —
> You are receiving this because you were mentioned.
>
>
> Reply to this email directly or view it on GitHub
> <https://github.com/facebook/folly/pull/378#issuecomment-209473788>
>