@nbronson 2016-04-13
I was hoping to reproduce the case that mattered to you. In this code readIndex_ and writeIndex_ are always accessed near each other, so I was going to see if colocating them on a cache line gave the same benefit with a smaller space penalty.