@rigtorp 2016-04-13
If readIndex_ and writeIndex_ are colocated on the same cache line it's
exclusive status will have to ping-pong between the reading and writing
cores. This causes memory bus level synchronization of reader and writer,
needlessly lowering performance.

On Wed, Apr 13, 2016 at 9:48 AM Nathan Grasso Bronson <
notifications@github.com> wrote:

> I was hoping to reproduce the case that mattered to you. In this code
> readIndex_ and writeIndex_ are always accessed near each other, so I was
> going to see if colocating them on a cache line gave the same benefit with
> a smaller space penalty.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub
> <https://github.com/facebook/folly/pull/378#issuecomment-209484275>
>