@nbronson 2016-04-13
If readIndex_ and writeIndex_ are separated then there will be two ping-pongs per enqueue or dequeue. If readIndex_ and writeIndex_ are colocated there could be less total ping-ponging, because a miss for one causes both to be filled. On the other hand, two independent misses can be overlapped, so maybe this would be slower even though it has fewer cache misses. Modern microarchitectures are very sophisticated and non-linear, so I'd like to test it on real hardware. It seems that the interleavings mainly depending on what fraction of time the program spends accessing the queue and the "heft" of T (the cost of T construction, access, and destruction, which goes in between the first and last access to the indexes during enqueuing and dequeuing). If you can give me an idea of your use case then I can make sure that I test it.