@Hincoin 2016-04-14
I ran some performance tests on an Intel Xeon E3-1505M v5 MB back when @rigtorp first submitted this PR and recently that tested the runtime of `ProducerConsumerQueue` across various high-utilization tests both with and without the cacheline separation on `writeIndex_` and `readIndex_`. 

On 1000 trials running each version of `ProducerConsumerQueue`, I found @rigtorp's proposal to be significant at the t = 4.786 level, which is a p-value of 1.163 * 10^-6, virtually zero. This makes sense too, since the separation would allow for various parallelization opportunities such as permitting a consumer to start reading from `readIndex_` while a producer is concurrently storing to `writeIndex_` and vice-versa. Further, it would allow for a concurrent producer and consumer to store to their respective `Index_` variables without worrying about ping-ponging the cacheline, among others. I would imagine that colocating the two variables upon the same cacheline prevents these sorts of opportunities for performance gains.

However, to quote Walter Bright, " measuring gives you a leg up on experts who are too good to measure.", so I'll run some tests to measure the difference in cache misses for each version and will post results back here later.