@rigtorp 2016-04-14
@nbronson You are mistaken in how cache coherency works: https://software.intel.com/en-us/articles/avoiding-and-identifying-false-sharing-among-threads and https://en.wikipedia.org/wiki/MOESI_protocol

@Hincoin It's good to measure, but I've implemented and benchmarked this so many times already. I added a figure to show all the false sharing mitigation present in the implementation I use https://github.com/rigtorp/SPSCQueue