@rigtorp 2016-04-14
@nbronson Also good resource: https://www.youtube.com/watch?v=WDIkqP4JbkE (slides http://www.aristeia.com/TalkNotes/codedive-CPUCachesHandouts.pdf)