@elliottneilclark 2016-07-06
What benchmarks are you using to test this? Can you include them with the PR?

For the existing benchmark that's in the folly directory I see this:


# Before Applying the PR
```
7	8	28127	196889
8	9	71014	568112
9	10	482	4338
10	11	55	550
11	12	156	1716
12	13	117	1404
13	14	38	494
14	15	10	140
16	17	1	16
============================================================================
folly/test/ProducerConsumerQueueBenchmark.cpp   relative  time/iter  iters/s
============================================================================
----------------------------------------------------------------------------
BM_ProducerConsumer(1048574)                               101.14ns    9.89M
BM_ProducerConsumerAffinity(1048574)                        89.32ns   11.20M
BM_ProducerConsumerLatency(1048574)                         1.67min   10.00m
============================================================================
```

# After Applying the PR

```
7	8	62230	435610
8	9	37130	297040
9	10	250	2250
10	11	60	600
11	12	125	1375
12	13	88	1056
13	14	96	1248
14	15	18	252
15	16	2	30
16	17	1	16
============================================================================
folly/test/ProducerConsumerQueueBenchmark.cpp   relative  time/iter  iters/s
============================================================================
----------------------------------------------------------------------------
BM_ProducerConsumer(1048574)                               100.29ns    9.97M
BM_ProducerConsumerAffinity(1048574)                       106.94ns    9.35M
BM_ProducerConsumerLatency(1048574)                         1.67min   10.00m
============================================================================
```

So from what I see this doesn't seem to make anything faster. It makes the case where threads have affinity worse.